parseint changed to parsefloat

+++ main.js
@@ -2,7 +2,7 @@ function add(number1, number2){
 
 
 
-	return parseInt(number1) + parseInt(number2);
+	return parseFloat(number1) + parseFloat(number2);
 
 }
 
@@ -10,7 +10,7 @@ function subtract(number1, number2){
 
 
 
-	return parseInt(number1) - parseInt(number2);
+	return parseFloat(number1) - parseFloat(number2);
 
 }
 
@@ -18,7 +18,7 @@ function multiply(number1, number2){
 
 
 
-	return parseInt(number1) * parseInt(number2);
+	return parseFloat(number1) * parseFloat(number2);
 
 }
 
@@ -26,7 +26,7 @@ function divide(number1, number2){
 
 
 
-	return parseInt(number1) / parseInt(number2);
+	return parseFloat(number1) / parseFloat(number2);
 
 }
